poc: smartpause: use wayland ext-idle-notify-v1 protocol
this uses maturin to add a (hacky) rust-based extension providing
bindings to ext-idle-notify-v1, and uses it on wayland (except gnome and
sway).
This has the downside of relying on (reverse-engineered, non-portable)
implementation details of pyo3/python-gi, and also introducing another
build step to safeeyes.
If this is to be actually used, it should a) be cleaned up, possibly by
splitting the rust extension into a separate package, and b) audited to
ensure that the unsafe code is actually correct, and possibly find some
other way to extract the wayland connection from gtk.